scripts/analysis/requirements.txt
# Data processing and handling
pandas
requests

# Snowflake connection
snowflake-connector-python[pandas]
snowflake-snowpark-python

# NLP and machine learning
vaderSentiment
transformers
scikit-learn
nltk

# Visualization and analysis
seaborn
matplotlib
scipy

# Environment management
python-dotenv
